projects/telegram-shop-orbit/requirements.txt
python-telegram-bot>=21.0
fastapi>=0.116.0
uvicorn>=0.35.0
